# docs/api-contract/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "KnowledgeScoreResponse": {
        "properties": {
          "articles_passed": {
            "title": "Articles Passed",
            "type": "integer"
          },
          "first_try_passes": {
            "title": "First Try Passes",
            "type": "integer"
          },
          "score": {
            "title": "Score",
            "type": "integer"
          },
          "user_id": {
            "title": "User Id",
            "type": "integer"
          }
        },
        "required": [
          "user_id",
          "score",
          "articles_passed",
          "first_try_passes"
        ],
        "title": "KnowledgeScoreResponse",
        "type": "object"
      },
...
      "QuizAdminUpdate": {
        "description": "Request body for the full-replace admin quiz save.",
        "properties": {
          "questions": {
            "items": {
              "$ref": "#/components/schemas/QuizQuestionAdmin"
            },
            "title": "Questions",
            "type": "array"
          }
        },
        "required": [
          "questions"
        ],
        "title": "QuizAdminUpdate",
        "type": "object"
      },
      "QuizAnswerAdmin": {
        "description": "Admin-facing candidate answer — includes is_correct.",
        "properties": {
          "answer_text": {
            "title": "Answer Text",
            "type": "string"
          },
          "id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Id"
          },
          "is_correct": {
            "default": false,
            "title": "Is Correct",
            "type": "boolean"
          },
          "position": {
            "default": 0,
            "title": "Position",
            "type": "integer"
          }
        },
        "required": [
          "answer_text"
        ],
        "title": "QuizAnswerAdmin",
        "type": "object"
      },
      "QuizAnswerPublic": {
        "description": "Operator-facing candidate answer — no is_correct.",
        "properties": {
          "answer_text": {
            "title": "Answer Text",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "answer_text"
        ],
        "title": "QuizAnswerPublic",
        "type": "object"
      },
      "QuizAttemptResult": {
        "properties": {
          "attempt_number": {
            "title": "Attempt Number",
            "type": "integer"
          },
          "passed": {
            "title": "Passed",
            "type": "boolean"
          },
          "score": {
            "title": "Score",
            "type": "integer"
          },
          "total_questions": {
            "title": "Total Questions",
            "type": "integer"
          },
          "wrong_question_ids": {
            "items": {
              "type": "integer"
            },
            "title": "Wrong Question Ids",
            "type": "array"
          }
        },
        "required": [
          "passed",
          "score",
          "total_questions",
          "wrong_question_ids",
          "attempt_number"
        ],
        "title": "QuizAttemptResult",
        "type": "object"
      },
      "QuizAttemptSubmit": {
        "description": "Submitted answers: {question_id: chosen_answer_id}.",
        "properties": {
          "answers": {
            "additionalProperties": {
              "type": "integer"
            },
            "title": "Answers",
            "type": "object"
          }
        },
        "required": [
          "answers"
        ],
        "title": "QuizAttemptSubmit",
        "type": "object"
      },
      "QuizPublicResponse": {
        "properties": {
          "article_id": {
            "title": "Article Id",
            "type": "integer"
          },
          "article_version": {
            "title": "Article Version",
            "type": "integer"
          },
          "questions": {
            "items": {
              "$ref": "#/components/schemas/QuizQuestionPublic"
            },
            "title": "Questions",
            "type": "array"
          }
        },
        "required": [
          "article_id",
          "article_version",
          "questions"
        ],
        "title": "QuizPublicResponse",
        "type": "object"
      },
      "QuizQuestionAdmin": {
        "description": "Admin-facing question — includes each answer's is_correct flag.",
        "properties": {
          "answers": {
            "items": {
              "$ref": "#/components/schemas/QuizAnswerAdmin"
            },
            "title": "Answers",
            "type": "array"
          },
          "id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Id"
          },
          "position": {
            "default": 0,
            "title": "Position",
            "type": "integer"
          },
          "question_text": {
            "title": "Question Text",
            "type": "string"
          }
        },
        "required": [
          "question_text",
          "answers"
        ],
        "title": "QuizQuestionAdmin",
        "type": "object"
      },
      "QuizQuestionPublic": {
        "description": "Operator-facing question — no is_correct anywhere in the payload.",
        "properties": {
          "answers": {
            "items": {
              "$ref": "#/components/schemas/QuizAnswerPublic"
            },
            "title": "Answers",
            "type": "array"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "question_text": {
            "title": "Question Text",
            "type": "string"
          }
        },
        "required": [
          "id",
          "question_text",
          "answers"
        ],
        "title": "QuizQuestionPublic",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/auth/login"
          }
        },
        "type": "oauth2"
      }
    }
  },
  "info": {
    "title": "Magti Internal Portal API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/articles/{article_id}/quiz": {
      "get": {
        "description": "Operator-facing quiz questions — no is_correct anywhere in the payload.",
        "operationId": "get_article_quiz_api_articles__article_id__quiz_get",
        "parameters": [
          {
            "in": "path",
            "name": "article_id",
            "required": true,
            "schema": {
              "title": "Article Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuizPublicResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Get Article Quiz",
        "tags": [
          "articles"
        ]
      }
    },
    "/api/articles/{article_id}/quiz/admin": {
      "get": {
        "description": "Full question set INCLUDING is_correct, for re-populating the admin\nquestion-builder when reopening an article for edit.",
        "operationId": "get_article_quiz_admin_api_articles__article_id__quiz_admin_get",
        "parameters": [
          {
            "in": "path",
            "name": "article_id",
            "required": true,
            "schema": {
              "title": "Article Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuizAdminUpdate"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Get Article Quiz Admin",
        "tags": [
          "articles"
        ]
      },
      "put": {
        "description": "Full delete-and-recreate of an article's quiz questions from the payload.",
        "operationId": "update_article_quiz_admin_api_articles__article_id__quiz_admin_put",
        "parameters": [
          {
            "in": "path",
            "name": "article_id",
            "required": true,
            "schema": {
              "title": "Article Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QuizAdminUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuizAdminUpdate"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Update Article Quiz Admin",
        "tags": [
          "articles"
        ]
      }
    },
    "/api/articles/{article_id}/quiz/attempt": {
      "post": {
        "description": "Grades a quiz submission server-side, records the attempt (pass or fail),\nand returns which questions were wrong so the operator can review before retrying.",
        "operationId": "submit_article_quiz_attempt_api_articles__article_id__quiz_attempt_post",
        "parameters": [
          {
            "in": "path",
            "name": "article_id",
            "required": true,
            "schema": {
              "title": "Article Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QuizAttemptSubmit"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuizAttemptResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Submit Article Quiz Attempt",
        "tags": [
          "articles"
        ]
      }
    },
...
    "/api/users/me/knowledge-score": {
      "get": {
        "description": "'ცოდნის ქულა' — +10 per distinct article quiz passed, +5 bonus if passed\non the very first attempt for that article version. Tunable constants, not\narchitecturally load-bearing.",
        "operationId": "get_my_knowledge_score_api_users_me_knowledge_score_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KnowledgeScoreResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Get My Knowledge Score",
        "tags": [
          "articles"
        ]
      }
    },
...
  }
}